+++ metrics/admin.py
@@ -56,7 +56,18 @@ class RequestAdmin(admin.ModelAdmin):
     fieldsets = (
         (
             _("Request"),
-            {"fields": ("method", "path", "full_path", "_query_string", "timestamp", "is_secure", "_headers")},
+            {
+                "fields": (
+                    "method",
+                    "path",
+                    "full_path",
+                    "_query_string",
+                    "timestamp",
+                    "is_secure",
+                    "is_ajax",
+                    "_headers",
+                )
+            },
         ),
         (_("Response"), {"fields": ("status_code",)}),
         (_("User info"), {"fields": ("referer", "user_agent", "ip", "_user", "language")}),
@@ -69,6 +80,7 @@ class RequestAdmin(admin.ModelAdmin):
         "_query_string",
         "timestamp",
         "is_secure",
+        "is_ajax",
         "_headers",
         "status_code",
         "referer",